Remove unused `chilldkg` DKG implementation
The deleted code included an outdated and unused Distributed Key Generation (DKG) implementation and associated tests. This cleanup eliminates unnecessary complexity and simplifies the codebase, ensuring only relevant components are retained.
