+++ src/common/constants.rs
@@ -1,0 +1,2 @@
+/// Timeout for DKG operations in seconds
+pub const DKG_TIMEOUT_SECONDS: u64 = 120;
