+++ src/common/types.rs
@@ -90,36 +90,20 @@ pub enum DkgMessage {
     Finish,
 }
 
-// /// Messages for distributed signing
-// #[derive(Debug, Clone, Serialize, Deserialize)]
-// pub enum SigningMessage {
-//     /// Start a new signing round for a message
-//     Start {
-//         /// The message to sign
-//         message: Vec<u8>,
-//         /// Participants involved in signing
-//         signers: Vec<Identifier>,
-//     },
-//     /// Nonce/commitment for round 1
-//     Round1 {
-//         /// Participant ID
-//         id: Identifier,
-//         /// Commitment data
-//         commitment: Vec<u8>,
-//     },
-//     /// Signature share for round 2
-//     Round2 {
-//         /// Participant ID
-//         id: Identifier,
-//         /// Signature share data
-//         signature_share: Vec<u8>,
-//     },
-//     /// Finalize signing by aggregating signature shares
-//     Finalize {
-//         /// All signature shares to aggregate
-//         shares: Vec<(Identifier, Vec<u8>)>,
-//     },
-// }
+/// State for a ChillDKG round
+#[derive(Debug, Clone, PartialOrd, PartialEq)]
+pub enum DkgRoundState {
+    /// Round 1: Generate and share commitments
+    Round1,
+    /// Round 2: Generate and exchange encrypted secret shares
+    Round2,
+    /// Round 3: Verify and finalize
+    Round3,
+    /// DKG complete
+    Complete,
+    /// DKG failed
+    Failed(String),
+}
 
 // Update SigningMessage enum to include the new message types
 /// Messages for distributed signing
