+++ src/lib.rs
@@ -2,5 +2,5 @@ pub mod common;
 pub mod frost;
 pub mod ipc;
 pub mod capnp_gen;
-mod chilldkg;
+pub mod chilldkg;
